feat(commands): from-pipe
